// freelancerprojectmanager.client/src/Model/PTask.ts
/**
 * corresponds to the dto returned by most apis
 */
export interface PTask {
   id: number;
  title: string;
  description: string;

  projectId: number;
  estimateMinute?: number;
  dueDate?: Date;
  plannedStart?: Date;
  status: PTaskStatus;
  createdAt: Date;
  completedAt?: Date;
}


export type PTaskStatus =
  | "ToDo"
  | "InProgress"
  | "Done"
  | "Canceled";